+++ packages/provider-utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/provider-utils
 
+## 3.0.0-alpha.11
+
+### Patch Changes
+
+- Updated dependencies [c1e6647]
+  - @ai-sdk/provider@2.0.0-alpha.11
+
 ## 3.0.0-alpha.10
 
 ### Patch Changes