+++ packages/xai/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ai-sdk/xai
 
+## 2.0.0-alpha.11
+
+### Patch Changes
+
+- Updated dependencies [c1e6647]
+  - @ai-sdk/provider@2.0.0-alpha.11
+  - @ai-sdk/openai-compatible@1.0.0-alpha.11
+  - @ai-sdk/provider-utils@3.0.0-alpha.11
+
 ## 2.0.0-alpha.10
 
 ### Patch Changes